+++ src/plugins/intel_cpu/src/nodes/executors/common/offset_helper.hpp
@@ -1,0 +1,80 @@
+// Copyright (C) 2018-2026 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include <bitset>
+
+#include "cpu_memory.h"
+#include "memory_desc/blocked_memory_desc.h"
+
+namespace ov::intel_cpu {
+class OffsetHelper {
+public:
+    static OffsetHelper createOffsetHelper(const MemoryPtr& mem) {
+        static const VectorDims empty_dims;
+        std::bitset<2> broadcast_mask;
+        if (nullptr == mem || mem->getDesc().empty()) {
+            return {nullptr, empty_dims, broadcast_mask, 0};
+        }
+        return createOffsetHelper(*mem);
+    }
+
+    static OffsetHelper createOffsetHelper(const IMemory& mem) {
+        std::bitset<2> broadcast_mask;
+        auto* base_ptr = static_cast<uint8_t*>(mem.getData());
+        auto desc = mem.getDescWithType<BlockedMemoryDesc>();
+        const auto& strides = desc->getStrides();
+        const auto prc = desc->getPrecision();
+        const auto& shape = desc->getShape().getStaticDims();
+        for (size_t i = 0; i < shape.size() && i < 2; i++) {
+            if (shape[i] == 1) {
+                broadcast_mask.set(i);
+            }
+        }
+        return {base_ptr, strides, broadcast_mask, prc.bitwidth()};
+    }
+
+    void* operator()(size_t i0) const {
+        if (!m_base_ptr) {
+            return nullptr;
+        }
+        if (m_broadcast_mask.test(0)) {
+            i0 = 0;
+        }
+        const size_t offset_bits = i0 * m_strides[0] * m_num_bits;
+        const size_t offset = div_up(offset_bits, 8);  // 8 bits in byte
+        return m_base_ptr + offset;
+    }
+
+    void* operator()(size_t i0, size_t i1) const {
+        if (!m_base_ptr) {
+            return nullptr;
+        }
+        if (m_broadcast_mask.test(0)) {
+            i0 = 0;
+        }
+        if (m_broadcast_mask.test(1)) {
+            i1 = 0;
+        }
+        const size_t offset_bits = i0 * m_strides[0] * m_num_bits + i1 * m_strides[1] * m_num_bits;
+        const size_t offset = div_up(offset_bits, 8);  // 8 bits in byte
+        return m_base_ptr + offset;
+    }
+
+    [[nodiscard]] void* get_base() const {
+        return m_base_ptr;
+    }
+
+private:
+    OffsetHelper(uint8_t* base_ptr, const VectorDims& strides, std::bitset<2> broadcast_mask, size_t num_bits)
+        : m_base_ptr(base_ptr),
+          m_strides(strides),
+          m_num_bits(num_bits),
+          m_broadcast_mask(broadcast_mask) {}
+
+    uint8_t* m_base_ptr = nullptr;
+    const VectorDims& m_strides;
+    size_t m_num_bits;
+    std::bitset<2> m_broadcast_mask;
+};
+};  // namespace ov::intel_cpu
