+++ src/plugins/intel_cpu/src/nodes/gathermatmul.cpp
@@ -231,7 +231,6 @@ void GatherMatmul::initSupportedPrimitiveDescriptors() {
                                                                       : MemoryDescUtils::makeEmptyDesc();
     descs[ARG_DST] = creatorsMap.at(LayoutType::ncsp)->createSharedDesc(dstTypes.front(), getOutputShapeAtPort(0));
 
-    m_attrs.isCompressedOperation = algorithm == Algorithm::GatherMatmulCompressed;
     auto executionContext = std::make_shared<ExecutorContext>(context, getImplPriority(), privateWeightCache);
     m_factory = std::make_shared<ExecutorFactory<GatherMatmulAttrs>>(m_attrs, executionContext, descs);
