+++ src/plugins/intel_cpu/src/transformations/cpu_opset/convert_to_cpu_specific_opset.hpp
@@ -44,9 +44,22 @@ inline void ConvertToCPUSpecificOpset(std::shared_ptr<ov::Model>& model, const C
     manager.set_per_pass_validation(false);
 
     // TransformMoeBlockToGatherMatmuls
-    CPU_REGISTER_PASS_COMMON(manager, ov::pass::ConvertTiledMoeBlockToGatherMatmuls);
-    CPU_REGISTER_PASS_COMMON(manager, ov::pass::Validate);
-    CPU_REGISTER_PASS_COMMON(
+    CPU_REGISTER_PASS_X64(manager, ov::pass::ConvertTiledMoeBlockToGatherMatmuls);
+    CPU_REGISTER_PASS_X64(manager, ov::pass::Validate);
+    CPU_REGISTER_PASS_X64(
+        manager,
+        ov::pass::ConvertGatherMatmulToGatherMatmulCompressed,
+        ov::intel_cpu::node::GatherMatmul::getSupportedCompressedActivationsTypes(),
+        ov::intel_cpu::node::GatherMatmul::getSupportedCompressedWeightsTypes(),
+        [&](const std::shared_ptr<ov::op::internal::GatherMatmulCompressed>& gather_matmul,
+            size_t IC,
+            size_t OC,
+            size_t G) {
+            return ov::intel_cpu::node::GatherMatmul::isSupportedCompressedOperation(gather_matmul, IC, OC, G, config);
+        });
+    CPU_REGISTER_PASS_ARM64(manager, ov::pass::ConvertTiledMoeBlockToGatherMatmuls);
+    CPU_REGISTER_PASS_ARM64(manager, ov::pass::Validate);
+    CPU_REGISTER_PASS_ARM64(
         manager,
         ov::pass::ConvertGatherMatmulToGatherMatmulCompressed,
         ov::intel_cpu::node::GatherMatmul::getSupportedCompressedActivationsTypes(),
